Initial run-time detection of C++ EH ABI variant.

+++ unwind-itanium.h
@@ -79,8 +79,8 @@ struct _Unwind_Exception
   {
     uint64_t exception_class;
     _Unwind_Exception_Cleanup_Fn exception_cleanup;
-    unsigned long private_1;
-    unsigned long private_2;
+    uintptr_t private_1;
+    uintptr_t private_2;
   } __attribute__((__aligned__));
 
 extern _Unwind_Reason_Code _Unwind_RaiseException (struct _Unwind_Exception *);
